# Conflicts:
#	WebVella.Erp.ConsoleApp/StringExtensions.cs
#	WebVella.Erp.Plugins.Crm/CrmPlugin._.cs
#	WebVella.Erp.Plugins.Mail/Api/AutoMapper/AutoMapperConfiguration.cs
#	WebVella.Erp.Plugins.Mail/Api/AutoMapper/SmtpServiceProfile.cs
#	WebVella.Erp.Plugins.Mail/Api/EmailServiceManager.cs
#	WebVella.Erp.Plugins.Mail/Api/SmtpService.cs
#	WebVella.Erp.Plugins.Mail/Hooks/Api/SmtpServiceRecordHook.cs
#	WebVella.Erp.Plugins.Mail/MailPlugin.20190215.cs
#	WebVella.Erp.Plugins.Mail/MailPlugin.20190420.cs
#	WebVella.Erp.Plugins.Mail/MailPlugin._.cs
#	WebVella.Erp.Plugins.Mail/MailPlugin.cs
#	WebVella.Erp.Plugins.Mail/Services/SmtpInternalService.cs
#	WebVella.Erp.Plugins.Mail/WebVella.Erp.Plugins.Mail.csproj
#	WebVella.Erp.Plugins.MicrosoftCDM/MicrosoftCDMPlugin._.cs
#	WebVella.Erp.Plugins.MicrosoftCDM/MicrosoftCDMPlugin.cs
#	WebVella.Erp.Plugins.MicrosoftCDM/WebVella.Erp.Plugins.MicrosoftCDM.csproj
#	WebVella.Erp.Plugins.Next/NextPlugin._.cs
#	WebVella.Erp.Plugins.Next/NextPlugin.cs
#	WebVella.Erp.Plugins.Next/Services/SearchService.cs
#	WebVella.Erp.Plugins.Project/Components/PcReportAccountMonthlyTimelogs/Display.cshtml
#	WebVella.Erp.Plugins.Project/Controllers/ProjectController.cs
#	WebVella.Erp.Plugins.Project/Files/timetrack.js
#	WebVella.Erp.Plugins.Project/ProjectPlugin._.cs
#	WebVella.Erp.Plugins.Project/ProjectPlugin.cs
#	WebVella.Erp.Plugins.Project/Services/CommentService.cs
#	WebVella.Erp.Plugins.Project/Services/FeedItemService.cs
#	WebVella.Erp.Plugins.Project/Services/TimeLogService.cs
#	WebVella.Erp.Plugins.Project/WebVella.Erp.Plugins.Project.csproj
#	WebVella.Erp.Plugins.SDK/Components/WvSdkPageSitemap/Form.cshtml
#	WebVella.Erp.Plugins.SDK/Components/WvSdkPageSitemap/form.js
#	WebVella.Erp.Plugins.SDK/Controllers/AdminController.cs
#	WebVella.Erp.Plugins.SDK/Pages/entity/list.cshtml
#	WebVella.Erp.Plugins.SDK/Pages/entity/relation-create.cshtml.cs
#	WebVella.Erp.Plugins.SDK/Pages/page/details.cshtml.cs
#	WebVella.Erp.Plugins.SDK/Pages/page/generated-body.cshtml.cs
#	WebVella.Erp.Plugins.SDK/SdkPlugin._.cs
#	WebVella.Erp.Plugins.SDK/SdkPlugin.cs
#	WebVella.Erp.Plugins.SDK/WebVella.Erp.Plugins.SDK.csproj
#	WebVella.Erp.Site.Crm/WebVella.Erp.Site.Crm.csproj
#	WebVella.Erp.Site.Mail/WebVella.Erp.Site.Mail.csproj
#	WebVella.Erp.Site.MicrosoftCDM/WebVella.Erp.Site.MicrosoftCDM.csproj
#	WebVella.Erp.Site.Next/WebVella.Erp.Site.Next.csproj
#	WebVella.Erp.Site.Project/WebVella.Erp.Site.Project.csproj
#	WebVella.Erp.Site.Sdk/WebVella.Erp.Site.Sdk.csproj
#	WebVella.Erp.Site/Startup.cs
#	WebVella.Erp.Site/WebVella.Erp.Site.csproj
#	WebVella.Erp.Web/Components/PcFieldAutonumber/PcFieldAutonumber.cs
#	WebVella.Erp.Web/Components/PcFieldCurrency/PcFieldCurrency.cs
#	WebVella.Erp.Web/Components/PcFieldDate/PcFieldDate.cs
#	WebVella.Erp.Web/Components/PcFieldEmail/PcFieldEmail.cs
#	WebVella.Erp.Web/Components/PcFieldNumber/PcFieldNumber.cs
#	WebVella.Erp.Web/Components/PcFieldPassword/PcFieldPassword.cs
#	WebVella.Erp.Web/Components/PcFieldPercent/PcFieldPercent.cs
#	WebVella.Erp.Web/Components/PcFieldPhone/PcFieldPhone.cs
#	WebVella.Erp.Web/Components/PcRow/Options.cshtml
#	WebVella.Erp.Web/Controllers/WebApiController.cs
#	WebVella.Erp.Web/ErpAppContext.cs
#	WebVella.Erp.Web/Middleware/ErpDebugLogMiddleware.cs
#	WebVella.Erp.Web/Models/Automapper/AutoMapperConfiguration.cs
#	WebVella.Erp.Web/Models/BaseErpPageModel.cs
#	WebVella.Erp.Web/Pages/HelloWorld.razor
#	WebVella.Erp.Web/Pages/RecordDetails.cshtml.cs
#	WebVella.Erp.Web/Repositories/SitemapAreaGroupRepository.cs
#	WebVella.Erp.Web/Services/AppService.cs
#	WebVella.Erp.Web/Services/CodeEvalService.cs
#	WebVella.Erp.Web/Services/PageService.cs
#	WebVella.Erp.Web/Services/UserFileService.cs
#	WebVella.Erp.Web/WebVella.Erp.Web.csproj
#	WebVella.Erp/Api/DataSourceManager.cs
#	WebVella.Erp/Api/EntityManager.cs
#	WebVella.Erp/Api/Models/AutoMapper/AutoMapperConfiguration.cs
#	WebVella.Erp/Api/Models/FieldTypes/BaseField.cs
#	WebVella.Erp/Database/DbConnection.cs
#	WebVella.Erp/ErpPlugin.cs
#	WebVella.Erp/Utilities/CryptoUtility.cs
#	WebVella.Erp/Utilities/Helpers.cs
#	WebVella.Erp/WebVella.Erp.csproj

+++ WebVella.Erp.Plugins.Mail/Api/AutoMapper/SmtpServiceProfile.cs
@@ -53,8 +53,8 @@ namespace WebVella.Erp.Web.Models.AutoMapper.Profiles
 			rec["default_sender_email"] = model.DefaultSenderEmail;
 			rec["default_sender_name"] = model.DefaultSenderName;
 			rec["default_reply_to_email"] = model.DefaultReplyToEmail;
-			rec["max_retries_count"] = (decimal)(model.MaxRetriesCount);
-			rec["retry_wait_minutes"] = (decimal)(model.RetryWaitMinutes);
+			rec["max_retries_count"] = (decimal)((int)model.MaxRetriesCount);
+			rec["retry_wait_minutes"] = (decimal)((int)model.RetryWaitMinutes);
 			rec["is_default"] = model.IsDefault;
 			rec["is_enabled"] = model.IsEnabled;
 			rec["connection_security"] = ((int)model.ConnectionSecurity).ToString();
